+++ classes/prefix_WPgutenberg/blocks/core-extensions/all/attributes.js
@@ -1,0 +1,12 @@
+const attributes = {
+  hideOnDesktop: {
+    type: 'boolean',
+    default: false,
+  },
+  hideOnMobile: {
+    type: 'boolean',
+    default: false,
+  }
+}
+
+export default attributes;