+++ classes/prefix_WPgutenberg/blocks/core-extensions/gallery/attributes.js
@@ -1,0 +1,16 @@
+const attributes = {
+  addSwiper: {
+    type: 'boolean',
+    default: false,
+  },
+  addPopUp: {
+    type: 'boolean',
+    default: false,
+  },
+  addPopUpPreview: {
+    type: 'boolean',
+    default: false,
+  }
+}
+
+export default attributes;