+++ classes/prefix_WPgutenberg/blocks/core-extensions/image/attributes.js
@@ -1,0 +1,8 @@
+const attributes = {
+  addPopUp: {
+    type: 'boolean',
+    default: false,
+  }
+}
+
+export default attributes;